refactor: axiosAPI baseURL 수정, fetchAPI 삭제
