+++ utils/axiosAPI/index.js
@@ -1,10 +1,9 @@
 import axios from "axios";
 
 const instance = axios.create({
+  baseURL: "http://localhost:3000/",
   timeout: 5000,
-  headers: {
-    "Cache-Control": "no-cache",
-  },
+  headers: {},
 });
 
 instance.interceptors.response.use(
